// frontend/src/providers/__generated__/TerraformProviderSearchFragment_providers.graphql.ts
export type TerraformProviderSearchFragment_providers$data = {
  readonly terraformProviders: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"TerraformProviderSearchListItemFragment_provider">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "TerraformProviderSearchFragment_providers";
};
export type TerraformProviderSearchFragment_providers$key = {
  readonly " $data"?: TerraformProviderSearchFragment_providers$data;
  readonly " $fragmentSpreads": FragmentRefs<"TerraformProviderSearchFragment_providers">;
};